third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-cors-redirect-expected.txt

CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php' (redirected from 'http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
[Worker] Test that basic EventSource cross-origin requests fail on redirect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-cors-redirect.js
PASS [Worker] es = new EventSource("/resources/redirect.php?code=307&url=http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php") did not throw exception.
PASS [Worker] es.readyState is EventSource.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
